versions: Cache whether versions are installed
The Versions.is_installed() function checks whether the executable exists on
the filesystem, and it is called for every version every time they are resorted
or refiltered, which includes every keystroke in the search bar. This is
inefficient and may make the interface stutter on lower-end computers.

Fixed by adding a simple cache.