ui: Use TabContainer in the main window
This makes the code simpler since we don't need to show/hide a bunch of nodes.
The TabContainer (with styling hidden) does that for us.

Also, make sure the tab buttons are updated when the tab is changed externally
(for example, when you click a project's version label, it takes you from
Projects to Versions).