versions tab: Add Cancel button to edit dialog
Changes are no longer applied immediately; you need to click Ok for them to
be saved. Also, if you create a new custom version, it won't be saved if you
change your mind and click Cancel. And it won't let you save the version if
the executable file is missing or the name is blank.